Fix final bug
